move normalization_utils
